UserRating 자료형을 Float 으로 수정

+++ 2004/nukeolaf/app/src/main/java/com/olaf/nukeolaf/Item.kt
@@ -8,5 +8,5 @@ data class Item(
     val pubDate: String,
     val director: String,
     val actor: String,
-    val userRating: Int
+    val userRating: Float
 )
